+++ Vault.js
@@ -1251,13 +1251,13 @@ class Vault {
   /**
    * @param {String<required>} token
    * @param {Object<required>} params
-   * @param {String} [params.kubernetes_host]
-   * @param {String} [params.kubernetes_ca_cert]
-   * @param {String} [params.token_reviewer_jwt]
+   * @param {String} params.kubernetes_host
+   * @param {String} params.kubernetes_ca_cert
+   * @param {String} params.token_reviewer_jwt
    * @param {Object} [params.pem_keys]
-   * @param {String} [params.issuer]
-   * @param {Boolean} [params.disable_iss_validation]
-   * @param {Boolean} [params.disable_local_ca_jwt]
+   * @param {String} params.issuer
+   * @param {Boolean} params.disable_iss_validation
+   * @param {Boolean} params.disable_local_ca_jwt
    * @param {String} mount
    * @returns {Object}
    */
@@ -1311,7 +1311,7 @@ class Vault {
    */
   async readK8sConfig(token, mount) {
 
-    assert(token, 'readK8sConfig: required parameter missing: vault token');
+    assert(token, 'readK8sConfig: required parameter missing: Vault admin token');
 
     let rootPath= "";
     if (mount) {
@@ -1341,25 +1341,24 @@ class Vault {
    * @param {String<required>} token
    * @param {String<required>} role
    * @param {Object<required>} params
-   * @param {String} [params.name]
    * @param {Object} [params.bound_service_account_names]
    * @param {Object} [params.bound_service_account_namespaces]
-   * @param {String} [params.audience]
-   * @param {Integer} [params.token_ttl]
-   * @param {Integer} [params.token_max_ttl]
+   * @param {String} params.audience
+   * @param {Integer or String} params.token_ttl
+   * @param {Integer or String} params.token_max_ttl
    * @param {Object} [params.token_policies]
    * @param {Object} [params.token_bound_cidrs]
-   * @param {Integer} [params.token_explicit_max_ttl]
-   * @param {Boolean} [params.token_no_default_policy]
-   * @param {Integer} [params.token_num_uses]
-   * @param {Integer} [params.token_period]
-   * @param {String} [params.token_type]
+   * @param {Integer or String} params.token_explicit_max_ttl
+   * @param {Boolean} params.token_no_default_policy
+   * @param {Integer} params.token_num_uses
+   * @param {Integer or String} params.token_period
+   * @param {String} params.token_type
    * @param {String} mount
    * @returns {Object}
    */
   async createK8sRole(token, role, params, mount) {
 
-    assert(token, 'createK8sRole: required parameter missing: vault token');
+    assert(token, 'createK8sRole: required parameter missing: Vault admin token');
     assert(role, 'createK8sRole: required parameter missing: role');
 
     // Defaults - most are probably already defaults from Vault itself
@@ -1368,6 +1367,10 @@ class Vault {
       token_num_uses: 0,
       ...params
     };
+
+    const { bound_service_account_names, bound_service_account_namespaces, audience,
+    token_ttl, token_max_ttl, token_policies, token_bound_cidrs, token_explicit_max_ttl,
+    token_no_default_policy, token_num_uses, token_period, token_type } = params;
 
     let rootPath= "";
     if (mount) {
@@ -1384,7 +1387,7 @@ class Vault {
         "X-Vault-Token": token
       },
       data: {
-        name, bound_service_account_names, bound_service_account_namespaces, audience,
+        bound_service_account_names, bound_service_account_namespaces, audience,
         token_ttl, token_max_ttl, token_policies, token_bound_cidrs, token_explicit_max_ttl,
         token_no_default_policy, token_num_uses, token_period, token_type
       }
@@ -1406,7 +1409,7 @@ class Vault {
    */
   async readK8sRole(token, role, mount) {
 
-    assert(token, 'readK8sRole: required parameter missing: vault token');
+    assert(token, 'readK8sRole: required parameter missing: Vault admin token');
 
     let rootPath= "";
     if (mount) {
@@ -1439,7 +1442,7 @@ class Vault {
    */
   async listK8sRoles(token, mount) {
 
-    assert(token, 'listK8sRoles: required parameter missing: vault token');
+    assert(token, 'listK8sRoles: required parameter missing: Vault admin token');
 
     let rootPath= "";
     if (mount) {
@@ -1473,7 +1476,7 @@ class Vault {
    */
   async deleteK8sRole(token, role, mount) {
 
-    assert(token, 'deleteK8sRole: required parameter missing: vault token');
+    assert(token, 'deleteK8sRole: required parameter missing: Vault admin token');
 
     let rootPath= "";
     if (mount) {