+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `0.4.4`
   * Added *Kubernetes (K8s) auth method* functions
+    * 
 
 * `0.4.3`
   * Added *KV v2 secret engine* function: